remove dependency from e2e/gardener on test/framework
